Splits library into specific concerns
